Add MarkdownService with extracted markdown functions.
