+++ server/src/services/pdf-service/PdfService.class.ts
@@ -4,11 +4,6 @@ import JSZip from 'jszip';
 import MarkdownIt from 'markdown-it';
 import path from 'path';
 import puppeteer from 'puppeteer';
-import {
-  PointMap,
-  convertExercisePointInfoToString,
-  ExercisePointInfo,
-} from 'shared/dist/model/Points';
 import { ScheincriteriaSummaryByStudents } from 'shared/dist/model/ScheinCriteria';
 import { Student } from 'shared/dist/model/Student';
 import { User } from 'shared/dist/model/User';
@@ -16,13 +11,13 @@ import { getNameOfEntity, sortByName } from 'shared/dist/util/helpers';
 import { getIdOfDocumentRef } from '../../helpers/documentHelpers';
 import Logger from '../../helpers/Logger';
 import { StudentDocument } from '../../model/documents/StudentDocument';
-import { TeamDocument } from '../../model/documents/TeamDocument';
 import { TutorialDocument } from '../../model/documents/TutorialDocument';
 import { BadRequestError, TemplatesNotFoundError } from '../../model/Errors';
+import markdownService, { TeamCommentData } from '../markdown-server/MarkdownService.class';
 import scheincriteriaService from '../scheincriteria-service/ScheincriteriaService.class';
 import sheetService from '../sheet-service/SheetService.class';
 import studentService from '../student-service/StudentService.class';
-import teamService, { PointInformation } from '../team-service/TeamService.class';
+import teamService from '../team-service/TeamService.class';
 import tutorialService from '../tutorial-service/TutorialService.class';
 import userService from '../user-service/UserService.class';
 import githubMarkdownCSS from './css/githubMarkdown';
@@ -30,16 +25,6 @@ import githubMarkdownCSS from './css/githubMarkdown';
 interface StudentData {
   matriculationNo: string;
   schein: string;
-}
-
-interface TeamCommentData {
-  teamName: string;
-  markdown: string;
-}
-
-interface SheetPointInfo {
-  achieved: number;
-  total: { must: number; bonus: number };
 }
 
 class PdfService {
@@ -96,7 +81,11 @@ class PdfService {
     teamId: string
   ): Promise<Buffer> {
     const [team] = await teamService.getDocumentWithId(tutorialId, teamId);
-    const { markdown } = await this.generateMarkdownFromTeamComment({ team, tutorialId, sheetId });
+    const { markdown } = await markdownService.generateMarkdownFromTeamComment({
+      team,
+      tutorialId,
+      sheetId,
+    });
 
     return this.generatePDFFromMarkdown(markdown);
   }
@@ -113,7 +102,7 @@ class PdfService {
 
     for (const team of tutorial.teams) {
       commentsByTeam.push(
-        await this.generateMarkdownFromTeamComment({ team, tutorialId, sheetId })
+        await markdownService.generateMarkdownFromTeamComment({ team, tutorialId, sheetId })
       );
     }
 
@@ -142,7 +131,7 @@ class PdfService {
   ): Promise<string> {
     const [team] = await teamService.getDocumentWithId(tutorialId, teamId);
 
-    const { markdown } = await this.generateMarkdownFromTeamComment({
+    const { markdown } = await markdownService.generateMarkdownFromTeamComment({
       team,
       sheetId,
       tutorialId,
@@ -181,81 +170,6 @@ class PdfService {
     if (notFound.length > 0) {
       throw new TemplatesNotFoundError(notFound);
     }
-  }
-
-  private async generateMarkdownFromTeamComment({
-    team,
-    tutorialId,
-    sheetId,
-  }: {
-    team: TeamDocument;
-    tutorialId: string;
-    sheetId: string;
-  }): Promise<TeamCommentData> {
-    const entries = await teamService.getPoints(tutorialId, team.id, sheetId);
-    const students = await teamService.getStudentsOfTeam(team);
-    const teamName = students.map(s => s.lastname).join('');
-
-    return { teamName, markdown: this.generateMarkdownFromPointEntries(entries, teamName) };
-  }
-
-  private generateMarkdownFromPointEntries(
-    entries: PointInformation[],
-    nameOfEntity: string
-  ): string {
-    const pointInfo: SheetPointInfo = { achieved: 0, total: { must: 0, bonus: 0 } };
-    let exerciseMarkdown: string = '';
-
-    entries.forEach(({ exName, entry, exPoints, subexercises }) => {
-      const achievedPts = PointMap.getPointsOfEntry(entry);
-      const exMaxPoints: string = convertExercisePointInfoToString(exPoints);
-
-      pointInfo.achieved += achievedPts;
-      pointInfo.total.must += exPoints.must;
-      pointInfo.total.bonus += exPoints.bonus;
-
-      exerciseMarkdown += `## Aufgabe ${exName} [${achievedPts} / ${exMaxPoints}]\n\n`;
-
-      if (typeof entry.points === 'object' && subexercises.length > 0) {
-        const subExData: { name: string; achieved: number; total: ExercisePointInfo }[] = [];
-
-        for (const subEx of subexercises) {
-          const achievedPts = entry.points[subEx.id];
-          subExData.push({
-            name: subEx.exName,
-            achieved: achievedPts || 0,
-            total: subEx.exPoints,
-          });
-        }
-
-        let subexerciseMarkdown = '';
-
-        subExData.forEach(({ name }) => {
-          subexerciseMarkdown += `|${name}`;
-        });
-        subexerciseMarkdown += '|\n';
-
-        subExData.forEach(() => {
-          subexerciseMarkdown += '|:-----:';
-        });
-        subexerciseMarkdown += '|\n';
-
-        subExData.forEach(({ achieved, total }) => {
-          const totalString = convertExercisePointInfoToString(total);
-          subexerciseMarkdown += `|${achieved} / ${totalString}`;
-        });
-        subexerciseMarkdown += '|\n';
-
-        exerciseMarkdown += `${subexerciseMarkdown}\n\n`;
-      }
-
-      exerciseMarkdown += `${entry.comment}\n\n`;
-    });
-
-    const totalPointInfo = convertExercisePointInfoToString(pointInfo.total);
-    const markdown = `# ${nameOfEntity}\n\n**Gesamt: ${pointInfo.achieved} / ${totalPointInfo}**\n\n${exerciseMarkdown}`;
-
-    return markdown;
   }
 
   private async generatePDFFromMarkdown(markdown: string): Promise<Buffer> {
